+++ README.md
@@ -1,5 +1,3 @@
-![Typescript: orientação a objetos](https://imgur.com/9399vxd.png)
-
 # Typescript: orientação a objetos
 
 Este projeto faz parte da escola Front-end e tem como objetivo ensinar a transformar trechos de código originalmente cosntruídos em programação funcional para programação orientada a objetos, destacando as diferenças entre os dois paradigmas.
